# src/main/resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "womapi.wom.desa",
  "basePath": "/ocp/common/srv40290-1-0/rest",
  "tags": [
    {
      "name": "customer-insight-controller",
      "description": "Customer Insight Controller"
    },
    {
      "name": "health-check",
      "description": "Health Check"
    }
  ],
  "paths": {
    "/customerInsight/checkCustomerCreditRating": {
      "post": {
        "tags": [
          "customer-insight-controller"
        ],
        "summary": "requestCustomerInsight",
        "operationId": "requestAuthorizeCreditUsingPOST",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CustomerCreditRequest"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "schema": {
              "$ref": "#/definitions/ResponsePOST"
            }
          },
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/ResponseGET"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "codigoError": {
                  "type": "integer",
                  "example": 400
                },
                "descripcionError": {
                  "type": "string",
                  "example": "test"
                },
                "message": {
                  "type": "string",
                  "example": "test"
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
    "/customerInsight/checkCustomerCreditRating/{id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "type": "string",
            "required": true
          }
        ],
        "tags": [
          "customer-insight-controller"
        ],
        "summary": "queryAuthorizeCredit",
        "operationId": "queryAuthorizeCreditUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/ResponseGET"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Record not found"
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
    "/health": {
      "get": {
        "tags": [
          "health-check"
        ],
        "summary": "getHealthCheck",
        "operationId": "getHealthCheckUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseHealth"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    }
  },
  "definitions": {
    "CustomerCreditRequest": {
      "type": "object",
      "properties": {
        "source": {
          "$ref": "#/definitions/Source"
        },
        "addData": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ExtraAttribute"
          }
        },
        "party": {
          "$ref": "#/definitions/Party"
        },
        "partyAddData": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ExtraAttribute"
          }
        },
        "portability": {
          "$ref": "#/definitions/Portability"
        },
        "portabilityAddData": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ExtraAttribute"
          }
        },
        "sourceAddData": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ExtraAttribute"
          }
        }
      },
      "title": "CustomerCreditRequest"
    },
    "ExtraAttribute": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "value"
        },
        "value": {
          "type": "string",
          "example": 15
        }
      },
      "title": "ExtraAttribute"
    },
    "Link": {
      "type": "object",
      "properties": {
        "href": {
          "type": "string"
        },
        "templated": {
          "type": "boolean"
        }
      },
      "title": "Link"
    },
    "Map��string,Link��": {
      "type": "object",
      "title": "Map��string,Link��",
      "additionalProperties": {
        "$ref": "#/definitions/Link"
      }
    },
    "ModelAndView": {
      "type": "object",
      "properties": {
        "empty": {
          "type": "boolean"
        },
        "model": {
          "type": "object"
        },
        "modelMap": {
          "type": "object",
          "additionalProperties": {
            "type": "object"
          }
        },
        "reference": {
          "type": "boolean"
        },
        "status": {
          "type": "string",
          "enum": [
            "100 CONTINUE",
            "101 SWITCHING_PROTOCOLS",
            "102 PROCESSING",
            "103 CHECKPOINT",
            "200 OK",
            "201 CREATED",
            "202 ACCEPTED",
            "203 NON_AUTHORITATIVE_INFORMATION",
            "204 NO_CONTENT",
            "205 RESET_CONTENT",
            "206 PARTIAL_CONTENT",
            "207 MULTI_STATUS",
            "208 ALREADY_REPORTED",
            "226 IM_USED",
            "300 MULTIPLE_CHOICES",
            "301 MOVED_PERMANENTLY",
            "302 FOUND",
            "302 MOVED_TEMPORARILY",
            "303 SEE_OTHER",
            "304 NOT_MODIFIED",
            "305 USE_PROXY",
            "307 TEMPORARY_REDIRECT",
            "308 PERMANENT_REDIRECT",
            "400 BAD_REQUEST",
            "401 UNAUTHORIZED",
            "402 PAYMENT_REQUIRED",
            "403 FORBIDDEN",
            "404 NOT_FOUND",
            "405 METHOD_NOT_ALLOWED",
            "406 NOT_ACCEPTABLE",
            "407 PROXY_AUTHENTICATION_REQUIRED",
            "408 REQUEST_TIMEOUT",
            "409 CONFLICT",
            "410 GONE",
            "411 LENGTH_REQUIRED",
            "412 PRECONDITION_FAILED",
            "413 PAYLOAD_TOO_LARGE",
            "413 REQUEST_ENTITY_TOO_LARGE",
            "414 URI_TOO_LONG",
            "414 REQUEST_URI_TOO_LONG",
            "415 UNSUPPORTED_MEDIA_TYPE",
            "416 REQUESTED_RANGE_NOT_SATISFIABLE",
            "417 EXPECTATION_FAILED",
            "418 I_AM_A_TEAPOT",
            "419 INSUFFICIENT_SPACE_ON_RESOURCE",
            "420 METHOD_FAILURE",
            "421 DESTINATION_LOCKED",
            "422 UNPROCESSABLE_ENTITY",
            "423 LOCKED",
            "424 FAILED_DEPENDENCY",
            "426 UPGRADE_REQUIRED",
            "428 PRECONDITION_REQUIRED",
            "429 TOO_MANY_REQUESTS",
            "431 REQUEST_HEADER_FIELDS_TOO_LARGE",
            "451 UNAVAILABLE_FOR_LEGAL_REASONS",
            "500 INTERNAL_SERVER_ERROR",
            "501 NOT_IMPLEMENTED",
            "502 BAD_GATEWAY",
            "503 SERVICE_UNAVAILABLE",
            "504 GATEWAY_TIMEOUT",
            "505 HTTP_VERSION_NOT_SUPPORTED",
            "506 VARIANT_ALSO_NEGOTIATES",
            "507 INSUFFICIENT_STORAGE",
            "508 LOOP_DETECTED",
            "509 BANDWIDTH_LIMIT_EXCEEDED",
            "510 NOT_EXTENDED",
            "511 NETWORK_AUTHENTICATION_REQUIRED"
          ]
        },
        "view": {
          "$ref": "#/definitions/View"
        },
        "viewName": {
          "type": "string"
        }
      },
      "title": "ModelAndView"
    },
    "Party": {
      "type": "object",
      "properties": {
        "currency": {
          "type": "string",
          "example": "CLP"
        },
        "nationalId": {
          "type": "string",
          "example": "23872783k"
        }
      },
      "title": "Party"
    },
    "Portability": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "string",
          "example": 9990
        },
        "operator": {
          "type": "string",
          "example": "Virgin Mobile"
        }
      },
      "title": "Portability"
    },
    "Source": {
      "type": "object",
      "properties": {
        "idApp": {
          "type": "string",
          "example": "WOMVOL"
        },
        "channel": {
          "type": "string",
          "example": "Web"
        },
        "user": {
          "type": "string",
          "example": "mcaror"
        },
        "branch": {
          "type": "string",
          "example": "Rosas"
        }
      },
      "title": "Portability"
    },
    "ResponsePOST": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "23872783k-1"
        },
        "status": {
          "type": "string",
          "example": "PROCESSING"
        }
      },
      "title": "ResponsePOST"
    },
    "ResponseGET": {
      "type": "object",
      "properties": {
        "authorizeCredit": {
          "type": "object",
          "properties": {
            "creditProfileDate": {
              "type": "string",
              "example": "2020-08-19T14:46:16.751-04:00"
            },
            "creditRiskRating": {
              "type": "string",
              "example": null
            },
            "idProfile": {
              "type": "string",
              "example": "PRE0015"
            },
            "attentionLevel": {
              "type": "string",
              "example": "P"
            },
            "@baseType": {
              "type": "string",
              "example": ""
            },
            "@schemaLocation": {
              "type": "string",
              "example": ""
            },
            "@type": {
              "type": "string",
              "example": ""
            }
          }
        },
        "creditScore": {
          "type": "object",
          "properties": {
            "product": {
              "type": "object",
              "properties": {
                "category": {
                  "type": "string",
                  "example": "OfertaGeneral"
                },
                "creditScore": {
                  "type": "integer",
                  "example": 0
                }
              }
            }
          }
        },
        "validFor": {
          "type": "object",
          "properties": {
            "startDateTime": {
              "type": "string",
              "example": "2020-08-19T14:46:16.751-04:00"
            },
            "endDateTime": {
              "type": "string",
              "example": null
            }
          }
        },
        "addData": {
          "type": "array",
          "items": {
            "oneOf": [
              {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "headerResponse.transaccion.internalCode"
                  },
                  "value": {
                    "type": "string",
                    "example": "2020012814613169"
                  }
                }
              },
              {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "headerResponse.transaccion.idTransactionBusiness"
                  },
                  "value": {
                    "type": "string",
                    "example": "2020012814613169"
                  }
                }
              },
              {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "OfertaComercial.oferta.tipoProducto"
                  },
                  "value": {
                    "type": "string",
                    "example": ""
                  }
                }
              },
              {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "OfertaComercial.oferta.cantidad"
                  },
                  "value": {
                    "type": "string",
                    "example": ""
                  }
                }
              },
              {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "OfertaComercial.oferta.script"
                  },
                  "value": {
                    "type": "string",
                    "example": "Cliente sin credito - intenta por otra vertical0"
                  }
                }
              }
            ]
          }
        }
      },
      "title": "ResponseGET"
    },
    "ResponseHealth": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "ok"
        }
      },
      "title": "ResponseHealth"
    },
    "View": {
      "type": "object",
      "properties": {
        "contentType": {
          "type": "string"
        }
      },
      "title": "View"
    }
  }
}